Run `synaptic_strength_fitting` test early in test suite
`simrun.synaptic_strength_fitting` Manages computation of voltage traces by computing a delayed as soon as the class is initialized. Only when you actually need the voltage, the result is collected/waited for.

While this is good practice, it clashes with the test suite, which imposes a strong time limit on how long tests are allowed to run. If the `PSPs` class is initialized at a "wrong" moment, it is possible the computation of voltage traces is scheduled on a busy time, in which case actually fetching the result may take a while, depending on which computations of other tests need to finish sooner.

This PR simply schedules the test earlier so that the scheduler queue is still empty when this test runs.