Make signature handles reflinks
Signature handles are e.g. def a.b.c.func
previously, these were rendered as plaintext, and a breadcrumb was shown
as a separate object above
now, the signature handle itself is rendered as internal reflinks,
effectively acting as a breadcrumb.
Added some css to make it look like the original signature handle.
unsure if this is a good idea or not (links should maybe look
likelinks?)