fix (?) filtering parameterfiles by content
Filtering parameterfiles by content is wise for speedup and ensuring you
only get unique parameterfiles, but it only really makes sense if you do
"hash_rename" copy method for parameterfiles. If you use "remount", you
want parameterfiles to preserve their different filenames, independent
of whether or not they may be identical in content. Otherwise, rerun_db
throws an error when it reruns e.g. a sim_trial_index that has a
paramfile with mismatching filename (but identical content) to the
paramfile it is looking for.